upload_file_running_on_VM_V1.1/requirements.txt
# Function dependencies, for example:
# package>=version
Flask==2.0.3
google-cloud-storage==1.42.3
Werkzeug==2.0.3
gunicorn
eventlet
